1 dia y 8 horas para el dia final

+++ src/hooks/useDeleteRedSocial.js
@@ -2,19 +2,29 @@ import axios from 'axios';
 import { API_BASE_URL } from '../markay/api/endpoint';
 
 const useDeleteRedSocial = () => {
-    
     const deleteData = async (id) => {
         try {
+            
+            const confirmDelete = window.confirm('¿Seguro que quieres eliminar esta red social?');
+            if (!confirmDelete) {
+                return; 
+            }
+
             const response = await axios.delete(`${API_BASE_URL}/api/v1/web/redSocial/${id}/`);
             if (!response.status.toString().startsWith('2')) {
                 throw new Error(`Error: ${response.status} ${response.statusText}`);
             }
+
             console.log('Datos eliminados con éxito:', response.data);
+
+
+      
         } catch (error) {
             console.error('Error al eliminar datos:', error.response ? error.response.data : error.message);
-            throw error; 
+            throw error;
         }
     };
+
     return deleteData;
 };
